Simplify the feature filter models.
